requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
diff-match-patch==20230430
Django==4.2.14
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-import-export==4.1.1
django-js-asset==2.2.0
django-mpesa==2.0.14
djangorestframework==3.15.2
idna==3.7
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
